pyproject.toml
[project]
name = "cron-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.3.2",
]
